Continuous release

+++ src/main/kotlin/de/livepoll/api/entity/dto/PollItemDtoOut.kt
@@ -1,8 +1,8 @@
 package de.livepoll.api.entity.dto
 
 open class PollItemDtoOut(
-        val itemId: Int,
-        val pollId: Int,
+        val itemId: Long,
+        val pollId: Long,
         val question: String,
         val position: Int,
         val type: String
